# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  response_entities.GenericResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: API TASK LIST
  title: github.com/brutalzinn/api-task-list
  version: "1.0"
paths:
...
  /apikey/revoke/{id}:
    delete:
      consumes:
      - application/json
      description: Revoke a user apikey
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response_entities.GenericResponse'
      summary: Revoke apikey
      tags:
      - ApiKeys
...
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task for current user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response_entities.GenericResponse'
      summary: Delete a task
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Get task by id for current user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response_entities.GenericResponse'
      summary: Get task by id
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Replace all tasks for a repo
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response_entities.GenericResponse'
      summary: Replace all tasks
      tags:
      - Tasks
...
swagger: "2.0"